Refactor: useInputs hook 추가
